Move parsing modules into their own directory
