+++ lib/content.js
@@ -1,6 +1,6 @@
 var logger = require('./logger');
 var lodash = require('lodash');
-var urlParser = require('./url-parser');
+var urlParser = require('./parse/url');
 var specSchema= require('./spec-schema');
 var contentTypeChecker = require('./content-type');
