+++ lib/middleware/route-map.js
@@ -1,6 +1,6 @@
 var glob = require('glob');
 var async = require('async');
-var parseBlueprint = require('../parse-blueprint');
+var parseBlueprint = require('../parse/blueprint');
 
 module.exports = function(options, cb) {
     var sourceFiles = options.sourceFiles;
