+++ lib/parse/action.js
@@ -1,5 +1,5 @@
-var urlParser = require('./url-parser');
-var route = require('./route');
+var urlParser = require('./url');
+var route = require('../route');
 
 module.exports = function(uriTemplate, action, routeMap) {
     var parsedUrl = urlParser.parse(uriTemplate);
