+++ lib/parse/blueprint.js
@@ -1,8 +1,8 @@
 var fs = require('fs');
 var protagonist = require('protagonist');
 var _ = require('lodash');
-var parseAction = require('./parse-action');
-var autoOptionsAction = require('./json/auto-options-action.json');
+var parseAction = require('./action');
+var autoOptionsAction = require('../json/auto-options-action.json');
 
 module.exports = function(filePath, autoOptions, routeMap) {
     return function(cb) {
