+++ lib/route.js
@@ -1,6 +1,6 @@
 var logger = require('./logger');
 var content = require('./content');
-var specSchema= require('./spec-schema');
+var specSchema = require('./spec-schema');
 
 exports.getRouteHandlers = function (method, parsedUrl, action) {
 
@@ -19,7 +19,7 @@ exports.getRouteHandlers = function (method, parsedUrl, action) {
                 }
             };
 
-            var matchRequests= function (specPair){
+            var matchRequests = function (specPair){
                 if (content.matches(req, specPair)) {
                     logger.log('[DRAKOV]'.red, action.method.green, parsedUrl.uriTemplate.yellow, (specPair.request && specPair.request.description ? specPair.request.description : action.name).blue);
